requirements.txt
aiohttp (3.6.3)
async-timeout (3.0.1)
attrs (20.3.0)
beautifulsoup4 (4.9.3)
chardet (3.0.4)
discord (1.0.1)
discord.py (1.5.1)
google (3.0.0)
idna (2.10)
idna-ssl (1.1.0)
multidict (4.7.6)
pip (9.0.1)
pkg-resources (0.0.0)
PyYAML (5.3.1)
setuptools (39.0.1)
soupsieve (2.0.1)
typing-extensions (3.7.4.3)
yarl (1.5.1)
